+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - implement validation for models
 - wrap `sqlalchemy.orm.relationship` with the configured `Query` class
 - override the `alembic` command to customize the generated migrations templates
+- export `ColumnMetaOption` from the top-level `sqlalchemy_unchained` package
 
 ## v0.1.0 (2018/09/24)