+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - wrap `sqlalchemy.orm.relationship` with the configured `Query` class
 - override the `alembic` command to customize the generated migrations templates
 - export `ColumnMetaOption` from the top-level `sqlalchemy_unchained` package
+- add documentation
 
 ## v0.1.0 (2018/09/24)