+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - export `ColumnMetaOption` from the top-level `sqlalchemy_unchained` package
 - export `association_proxy`, `declared_attr`, `hybrid_method`, and `hybrid_property` from the top-level `sqlalchemy_unchained` package
 - add documentation
+- add tests
 
 ## v0.1.0 (2018/09/24)